step1 move minesposition generate function
